+++ Examples/src/main/java/org/openestate/io/examples/CasaItWritingExample.java
@@ -23,7 +23,6 @@ import java.math.BigDecimal;
 import java.math.BigInteger;
 import org.apache.commons.io.output.NullOutputStream;
 import org.apache.commons.io.output.NullWriter;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.RandomUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.SystemUtils;