+++ Examples/src/main/java/org/openestate/io/examples/ImmobiliareItWritingExample.java
@@ -25,7 +25,6 @@ import java.util.Calendar;
 import java.util.Currency;
 import org.apache.commons.io.output.NullOutputStream;
 import org.apache.commons.io.output.NullWriter;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.RandomUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.SystemUtils;