+++ Examples/src/main/java/org/openestate/io/examples/OpenImmoWritingExample.java
@@ -24,7 +24,6 @@ import java.util.Calendar;
 import java.util.Locale;
 import org.apache.commons.io.output.NullOutputStream;
 import org.apache.commons.io.output.NullWriter;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.RandomUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.SystemUtils;