+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Changelog for OpenEstate-IO
 1.3 (not released yet)
 ----------------------
 
+-   Updated `commons-lang` library to version 3.6.
+-   Add dependency to `commons-text` library in version 1.1.
+
 
 1.2 (14 May 2017)
 -----------------